Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/css/cssom/CSSKeyframesRule-expected.txt


FAIL CSSOM -  CSSKeyframesRule interface assert_equals: CSSKeyframesRule cssText attribute expected "@keyframesbar{}" but got "@-webkit-keyframesbar{}"
